Correct "Operator" Auto Attendant default prompt.
The default prompt for the "Operator" Auto Attendant was being set to
"welcome.wav", which isn't the correct prompt--nor was that prompt stored 
in the correct place to be found for uploaded AA prompts.  This caused the 
AA to hangup on calls if the default configuration was selected.

Changed the default to "autoattendant.wav"

Fixes XPB-1004.

--Woof!


git-svn-id: https://sipxecs.sipfoundry.org/rep/sipXecs/main@10100 ab1d8caa-1f67-47f1-9e81-24633a41865c