Generic changes:
Make UtlRegex containable.


Proxy changes:

Add code to support 'routeIPv4subnet' and 'routeDnsWildcard' 
elements from forwardingrules.xml.  This allows routing based on
subnets and DNS wildcards, to enable direct routing to border elements and
ITSPs/foriegn proxies for URL dialed calls that are sent to this proxy (due
to default proxy settings on phones/UAs), but are not addressed to this
proxy.

Add optional boolean attribute 'authRequired' to 'routeTo' element in
forwardingrules.xml.  This enables the inserted route to also include
 a route to the authproxy first.  It is expected that ALL external calls
be routed via the authproxy, and this, combined with the above patterns,
enables this.


Auth Proxy changes:

Acutally make 'hostPattern' element in authrules.xml a pattern!  Previously
(despite the name) it was just a match on hostname and port.  Added
a new optional attribute 'format' which can have a value of "url", 
"IPv4subnet", or "DnsWildcard".  This allows those types of patterns to be 
matched for authorization just as the above allowed the proxy to route on
patterns.


Registrar changes:

Since 'hostPattern' is also used in fallbackrules.xml, and mappingrules.xml,
The rewritting rules can also take advantage of the new patterns.  Please note
that if the rewritten URL is external, the <transform> elements should include
<headerparams>route={extern}</headerparams> to tell the proxy to route the 
call to the external entitiy.  The proxy will automatically add a route to
the authproxy when it receives the redirect with a route parameter.


Patterns:

IPv4subnet matches IPv4 subnet, specified in modified CIDR notation as a 
series of IPv4 address decimal octet values separated by dots (.), followed 
by a forward slash (/) and a decimal number of bits to be used as a subnet mask
value.

No port may be specified, and port is not considered in the match.

Examples:
   10.0.0.0/8  Matchs any IP address with first 8 bits = '10'
   192.168/16  

Note that DNS names are NOT looked up: this pattern only works if the URL
contains an IP address, not a FQDN.


DnsWildcard matches "wildcard" DNS Domains, specified as a pattern  whose 
leftmost lable is '*'.  This part of the pattern matches any sequence of DNS
name tokens and dots.  The rest of the pattern must match the full domain
value from the request URI.  No port may be specified, and port
is not considered in the match.

Examples:
   *.example.com (matches any subdomain in example.com INCLUDING
                     example.com itself)
   * (matches all domains)

--Woof!


git-svn-id: https://sipxecs.sipfoundry.org/rep/sipXecs/main@10149 ab1d8caa-1f67-47f1-9e81-24633a41865c