Remove confusing EXTENSION_LENGTH setting from System->General->Voice Mail
This value was used to limit the digit entry in Voice Mail when the
system asked for an extension (such as when forwarding voice mail to some
other extension).

Users were confusing it with some Dial Plan digit lengths, and Auto Attendant
features (because the comment describing it incorrectly mentioned Auto Attendant).

As there is a short (3 second) timeout, and the user can always end the digit
entry with '#' to skip the timeout, changing this has little effect on the
voice mail user experience.

I changed it to a hard coded maximum length of 100, which should be large
enough even for the most hardened DTMF user.

--Woof!


git-svn-id: https://sipxecs.sipfoundry.org/rep/sipXecs/main@10177 ab1d8caa-1f67-47f1-9e81-24633a41865c