Fix XECS-64: Call pickup continues to route *78... to default GW if no phones report ringing
In a perfect world, the mapping rules would not match the call pickup
codes.  But often they do.  In the example in XECS-64, the pattern *781234
matched 7 digit dialing.  So the external local gateway was added as a
contact.  If the pickup code found a ringing phone, this didn't matter as
the pickup code ignored it.  However, if no ringing phone was found
(because there wasn't one, or the ringing phone didn't support the SUBSCRIBE,
or the ringing phone just didn't respond in time), then the previously
found contacts would be used, causing the phone that dialed the pickup code
to be directed out the gateway.

To fix this, once the pickup code is matched, any previously added contacts 
are incorrect, and no longer relevent, so remove any contacts added by 
previous redirectors before adding new contacts.  This way, if no contacts 
are added by the PickUp redirector, no contacts will be found and a 404 will 
be returned to the phone that dialed the pickup code.  This makes sense, 
as no busy phones were found.

Note:  This does not solve problems with contacts that are added AFTER
PickUp, but that depends on redirector order, and currently PickUp is 
at the end.

--Woof!


git-svn-id: https://sipxecs.sipfoundry.org/rep/sipXecs/main@10225 ab1d8caa-1f67-47f1-9e81-24633a41865c