Unescape audio file:// URLs when playing them.
Now that the config server correctly escapes file:// URLs, the path part
needs to be unescaped before opening that file.

Fixes XECS-149 (The background music file name created for sipXpark should 
not contain space in it since sipXpark does not handle it well.)

It's not sipXpark where the problem lay-- it was deep inside the sipXmediaLib.
This change should work for all audio file playing processes (ACD, sipXpark, 
sipXvxml).

--Woof!


git-svn-id: https://sipxecs.sipfoundry.org/rep/sipXecs/main@10232 ab1d8caa-1f67-47f1-9e81-24633a41865c