Add a unique "call-id" to the ACD Events log
Instead of relying on "from" to be unique, use the sipXtapi "callHandle" (an
ever-increasing integer) as the unique call identifyer.

callHandle will reset to 1 each time the ACD is started, but that is okay as
the calls cannot survive a restart anyway.

Includes ruby script changes from Damian to parse the new format.

Fixes XECS-230 - ACD real-time statistics depend on "From" field uniquely
identify the call

--Woof!


git-svn-id: https://sipxecs.sipfoundry.org/rep/sipXecs/main@10386 ab1d8caa-1f67-47f1-9e81-24633a41865c