Enable per-user voice mail VoiceXML scripts.
This is the start of allowing personalized voice mail, such as a different
operator for each user (XECS-186) and the ability to transfer voice mail calls
to several user specified destinations (XECS-185).

After several false starts, it turned out to be deceptively easy (on the server
side).  The hard part is still to come, which is config server generating the
appropriately customized scripts.  To that end, the existing voice mail
VoiceXML script needs to be re-factored and parametrized, as the Auto
Attendant script is today.

When a call arrives for voice mail (specifically, the media server is given a
URL that runs the DepositCGI script), that script will now look in the user's
voice mail box directory (/var/sipxdata/mediaserver/data/mailstore/{mailbox}/)
for a script named "savemessage.vxml".  If it exists, that is the script that
becomes the src URL for a subdialog.  If it doesn't exist, than the generic
script at http://{mediaserver}/vm_vxml/savemessage.vxml is invoked instead.

--Woof!


git-svn-id: https://sipxecs.sipfoundry.org/rep/sipXecs/main@10402 ab1d8caa-1f67-47f1-9e81-24633a41865c