Modified the interface to an AuthPlugin and the SipAaa logic that calls them.
The AuthResult type returned by an AuthPlugin is changed
  from:
       { ALLOW_REQUEST, UNAUTHORIZED }
  to:
       { CONTINUE, ALLOW, DENY }

Every AuthPlugin is now called for every message, but the final result of the
authorization decision is the first AuthResult that is not CONTINUE - so the
order of the plugins (which is controlled by the configuration) now determines
the precedence of the decisions made by different AuthPlugin modules.

The AuthPlugin::authorizeAndModify interface gets two additional arguments:

  AuthResult       priorResult

    Passing the priorResult (the result as determined so far) allows the
    AuthPlugin to do some optimization.  An AuthPlugin that makes an
    authorization decision can skip any of that processing if the
    priorResult is not CONTINUE, and an AuthPlugin that might modify the
    message can skip doing so if the priorResult is DENY.

  const UtlString& method

    The SIP method will be used by many AuthPlugins, and the
    SipAaa::proxyMessage method has already parsed it out, so passing it
    saves repeating the same processing in each AuthPlugin.

The SipAaa::proxyMessage method now directly checks two authorization
decisions: if the message is an ACK or a response, then it should
always be allowed.  This was implemented in EnforceAuthRules, but
should have been common to all authorization decisions so it has been
moved into the caller (and the priorResult is set to ALLOW before the
first AuthPlugin is called).

The names for the two existing default AuthPlugin modules have been
changed to explicitly control the order; a map has been created to 
change these automatically in an RPM build.



git-svn-id: https://sipxecs.sipfoundry.org/rep/sipXecs/main@10797 ab1d8caa-1f67-47f1-9e81-24633a41865c