XCF-1860 Allow specification of custom port and/or transport when creating/modifying gateways
There is new capability to specify custom port and transport protocol for the gateway. Both fields are "hidden" into advanced settings using ToggleLink.
If transport is set to UDP/TCP then only that particular protocol is used for outgoing requests towards the gateway. If the protocol is not specified DNS SRV query will determine the transport.
The default is set to UDP.


git-svn-id: https://sipxecs.sipfoundry.org/rep/sipXecs/main@10818 ab1d8caa-1f67-47f1-9e81-24633a41865c