XCF-1861 make detecting duplicate schedule names compatible with global schedules
Detecting schedule names duplicate was implemented before introduction of global schedules.
ForwardingContextImpl.isNameInUse was making a false assumption about what type od schedules are available.
I changed isNameInUse so that it only checks for schedules types that it knows about it.
This change does not actually implement this feature for global schedules, but it makes it possible for global schedules to be used until detecting duplicates is implemented.


git-svn-id: https://sipxecs.sipfoundry.org/rep/sipXecs/main@10838 ab1d8caa-1f67-47f1-9e81-24633a41865c