XCF-1839 change defaults for voicemailFallback property of the call group
To mirror 3.8 bahavior:
- for newly created call groups voicemailFallback is 'on'
- for existing callgroups (during upgrade) it is 'off' if alernative falback destination is set and 'on' otherwise
That should ensure that if hunt group was redirected to voicemail in 3.8 it will be redirected to voicemail after upgrade.

For technical reasons in 3.8 and before we could not refirect to voicemail calls groups with the parallel fork on last stage.
This fix removes such limitation: sipXconfig can generarate alias for last user voicemail as long as call group has at least one member.


git-svn-id: https://sipxecs.sipfoundry.org/rep/sipXecs/main@10947 ab1d8caa-1f67-47f1-9e81-24633a41865c