Enable RECOVERABLE_CRITICAL_SECTION in fastdb.
This forces the use of SysV semaphores for the critical section with 
SEM_UNDO, and avoids the "magic" check on a static variable using inline 
asm code that is rumored not to work with multiprocessor implementations.

In the older version of the fastdb, this section was avoided because
it left the SEM_UNDO counts in odd states.  The newer version of
fastdb removed the SEM_UNDO counts, but still uses the inline asm.

Best to be safe and use the old fashoned (but known to work) version.

--Woof!


git-svn-id: https://sipxecs.sipfoundry.org/rep/sipXecs/main@11286 ab1d8caa-1f67-47f1-9e81-24633a41865c