XCF-1832 LDAP preview: display error message instead of Internal exception
Catches the exceptions thrown by LdapTemplate's search method (of type org.springframework.dao.DataAccessException) and converts them in a UserException to be rendered in the web browser.

Please note that beginning from version 1.2 of spring-ldap library (we're now using 1.1.2), the search method will throw errors of type org.springframework.ldap.NamingException so when we'll upgrade to 1.2 we'll need to catch that instead of DataAccessException.

Patch from Mircea Amarascu - thanks!


git-svn-id: https://sipxecs.sipfoundry.org/rep/sipXecs/main@11359 ab1d8caa-1f67-47f1-9e81-24633a41865c