Add intrusive logging to fastdb
However...it will only log if the file "/tmp/wtf.log" exists
and is world writable.

This should be not so intrusive unless that file exists, so
it shouldn't effect anyone unless they are looking for 
trouble.

Logging output is placed in the /tmp/wtf.log.  Please create
the file with 0666 permissions so all users can write to it.
This command should suffice:

sudo touch /tmp/wtf.log; sudo chmod 666 /tmp/wtf.log

And yes, "wtf" stands for "what the fluff", as in
"What the fluff is going on in the imdb?"

--Woof!


git-svn-id: https://sipxecs.sipfoundry.org/rep/sipXecs/main@11429 ab1d8caa-1f67-47f1-9e81-24633a41865c