Add one single critical section around database calls.
One mutex to rule them all!

Uses flock() on the imdb.im file as a global process mutex,
combined with a local counter and mutex to enable thread re-entrant 
calls (so locks can be nested per thread), but not across threads.

flock() has the nice characteristic that it releases when the
file descriptor is closed, and so a crashed process also automatically
releases any held lock.

The reason for this is to attempt to avoid the fastdb internal
lock counters, which appear to have problems we have yet to fathom.
With one critical section around everything, the fastdb shouldn't
ever use its read/write semaphores.  The cost is concurrency.
Performance may suffer.

--Woof!


git-svn-id: https://sipxecs.sipfoundry.org/rep/sipXecs/main@11439 ab1d8caa-1f67-47f1-9e81-24633a41865c