Fix for XECS-657: Park server dialog events sometimes include incorrect state information
I believe this problem was introduced as part of the authenticated identities work.  Essentially, the subscribe server included the ~~id~park in its contact address which caused some clients (e.g. dialogwatch) to re-subscribe to ~~id~park instead of the desired resource-id.

Now, the subscribe server sets the userId in the contact to match that of resource id.

I also updated the subscribe contactTest unit tests.  These tests assumed that the contact would never include a userId.  Also, some of unit tests mismatched notify requests and responses in the message queues.



git-svn-id: https://sipxecs.sipfoundry.org/rep/sipXecs/main@11510 ab1d8caa-1f67-47f1-9e81-24633a41865c