XCF-2012 Improve deleting schedules.
Instead of showing warnings like "The following schedules were not deleted because they are associated with forwarding rules / dialing rules ...", the new behavior is to delete the schedule and switch all the associated resources (forwarding rules / dialing rules) to "Always" schedule.
The new behavior explained above applies for all kinds of schedules (general, user and group schedules)

This was done using a listener which is triggered when deleting a schedule. The listener checks for the schedule's type (general/user/group) and associates the corresponding forwarding/dialing rules with "Always" schedule.

Patch from Bogdan Brezoi - thanks!


git-svn-id: https://sipxecs.sipfoundry.org/rep/sipXecs/main@11522 ab1d8caa-1f67-47f1-9e81-24633a41865c