XCF-1896 handle gracefully cases when call resolver is not running
When call resolver is not running the active tab is hidden and the following error message is displayed:
WARNING: Unable to connect to Call Resolver interface. Call resolver service may not be installed.

The History tab is shown always: it does not require that call resolver is running

Patch from Mircea Carasel - thanks!


git-svn-id: https://sipxecs.sipfoundry.org/rep/sipXecs/main@11553 ab1d8caa-1f67-47f1-9e81-24633a41865c