XECS-742 increase frequency of flushing failed calls
Since detecting failed calls is difficult call resolver is marking calls as failed and flushes them periodically. In the passed the flushing frequency was based on number of events received since the call started. It did not work well if the call volume was low.
Now we periodically check the failed call cache and remove the calls that are older then SIP_CALLRESOLVER_MAX_FAILED_WAIT (default 3 minutes).


git-svn-id: https://sipxecs.sipfoundry.org/rep/sipXecs/main@11563 ab1d8caa-1f67-47f1-9e81-24633a41865c