XCF-1929 make session timeout adjustable from sipxconfig.properties
Increase default timeout to 1 hour for admin sessions and 30 minutes for user sessions (used to be 15 minutes for all sessions).
There is not UI for now but this value is configurable in sipxconfig.properties.in file
To adjust the timeout add the following lines:

tapestry.adminSessionTimeout=60
tapestry.userSessionTimeout=30

Timeout is expressed in minutes.

Based on the patch from Mircea Carasel - thanks!


git-svn-id: https://sipxecs.sipfoundry.org/rep/sipXecs/main@11569 ab1d8caa-1f67-47f1-9e81-24633a41865c