Add dump facility to sipXrls: Upon receiving a MESSAGE to URI "sip:*dumpstate@[domain]", the RLS will dump its state into the log. This can be done by executing: sipsend.pl [host]:5140 MESSAGE 'sip:*dumpstate@[host]:5140'
Add most of the dump code to SipSubscribeServer and
SipSubscribeClient.

Change the file name argument of ResourceListFileReader::_ from 'const
UtlString*' to 'const UtlString&', because the object does not retain
a pointer to the UtlString.  It also allows us to eliminate a compiler
warning.

Add ResourceListServer::getServerUserAgent().

Replace various UtlHashMap's with UtlHashBag's where that was what the
design intended.

Set the task name of ResourceListTask's.

Correct and expand various comments.  Improve formatting.  Delete
unused variables.


git-svn-id: https://sipxecs.sipfoundry.org/rep/sipXecs/main@11593 ab1d8caa-1f67-47f1-9e81-24633a41865c