XCF-1992/XCF-2060 diagnostic tests UI changes
All tests and test results are now displayed on a single page.
More information about the tests is displayed in case of failure/error or warning.
One can now style the diagnostic error status: class name is composed of the 'status-' prefix followed by the name of the enumeration value. status-InProgres, status-Error etc.


git-svn-id: https://sipxecs.sipfoundry.org/rep/sipXecs/main@11680 ab1d8caa-1f67-47f1-9e81-24633a41865c