Add ACD Agent automatic sign-out and RNA wrap up time.
Add the automatic sign-out of an agent that "Rejects"
or doesn't answer a call, after some number of attempts.
(Default is 3, defined as a per queue attribute "Maximum Bounce Count"

Also add a "Reject" and RNA "wrap up" time, so that the
agent is not available for a short period after she
"Rejects" or doesn't answer.  This lets the call be
routed to another, available agent.  Default is 30 seconds
as a per queue attribute "Agent Non-Responsive time".

Parts of this from the XECS-754 patch by Huijun Yang.

Also fixes issues in sipxpresence which would core
dump trying to shutdown http connections.

Fixes XECS-754 - ACD Agent automatic sign-out
Fixes XECS-749 - A caller hears clicking when an agent that is signed in has no phone

--Woof!


git-svn-id: https://sipxecs.sipfoundry.org/rep/sipXecs/main@11755 ab1d8caa-1f67-47f1-9e81-24633a41865c