XCF-2147.patch autorefresh for Acd presence forms
Change the classic form into an AutoRefreshForm for ACD Presence page, both for the superadmin user and the normal user.
ACD Presence page for the user changed: now both the sign in and sign out buttons are available. Before this change, only one button was present (depending on the signing status).
This change is necessary because AutoRefreshForm component does not refresh the action block which contains the button.

Patch from Bogdan Brezoi - thanks!


git-svn-id: https://sipxecs.sipfoundry.org/rep/sipXecs/main@11844 ab1d8caa-1f67-47f1-9e81-24633a41865c