XCF-2189 improvements for mrgt.conf parser
The parser now recognizes and ignores comments (line starting with #)
Also, the reason for the error described in this issue is not necessarily related to comments in files.
In this particular case, the error appeared because the MRTGContext:intializeConfigFiles() method was expecting, in the mrtg.cfg file, target id's like templateTargetId_host . In our exposed case, the target id had no "_" in it, so the code threw a StringIndexOutOfBoundsException.
This patch updates MRTGContextImpl to consider target ids without underscore as invalid targets and these targets will be deleted from the mrtg.cfg file.
Updates MRTGTarget.java - ignoring empty String.

Patch from Andrei Niculae - thanks!


git-svn-id: https://sipxecs.sipfoundry.org/rep/sipXecs/main@11871 ab1d8caa-1f67-47f1-9e81-24633a41865c