XCF-2196: Implemented variable-width sidebar layout.
Implemented LeftNavigation as a component. The left nav content goes here.
For this to work, the LeftNavigation component must be at the same level as the Border component or nested one level deep.

The LeftNavigation component supports an optional "renderCondition" parameter (a boolean value).
This will cause the left navigation links to be rendered only if the condition evaluates to true.
Pages/components that surrounded the LeftNavigation component with an If component have been changed to use
this new parameter.

The @If components were not working properly since they just control the rendering of their bodies, and the LeftNavigation component is not rendered where it appears in the page.

Patch from Joe Attardi. Thanks Joe!


git-svn-id: https://sipxecs.sipfoundry.org/rep/sipXecs/main@11902 ab1d8caa-1f67-47f1-9e81-24633a41865c