XECS-726 Remove volatile gateways when reseting dial plan.
3rd party plugins can modify dial plan by providing dial plan definion in XML spring bean file.
Gateways that are defined in such files have to be removed automatically when dial plan is reset (otherwise we are trying to reinsert them, which causes duplicate errors)
This patch introduces a special "volatile" marked in gateway model definition. Volatile gateways are removed from DB when dial plan is reset.


git-svn-id: https://sipxecs.sipfoundry.org/rep/sipXecs/main@11913 ab1d8caa-1f67-47f1-9e81-24633a41865c