XCF-2239: Fixed the Localization page so that the current language is properly selected in the Language drop-down.
Also fixed a couple of typos ("langauge", "optained").

On pageBeginRender, the current language was being set to the full language ID ("stdprompts_en") instead of the language code ("en"), which is what the drop-down needs to select the proper value.

So I just changed the pageBeginRender() method to use the value of localizationContext.getCurrentLanguage() instead of localizationContext.getCurrentLanguageDir().

Patch from Joe Attardi.


git-svn-id: https://sipxecs.sipfoundry.org/rep/sipXecs/main@11950 ab1d8caa-1f67-47f1-9e81-24633a41865c