Add ability to check imdb semaphore states in ipcclean.pl
Added a new argument --checksems to ipcclean.pl.  This will
attempt to aquire/check the various imdb semaphores.  If they 
are not in the correct state after waiting up to 60 seconds, it 
exits with an error.

This script can then easily be incorporated into a monitoring
system, via cron, "watch", or whatever to report or take action if the
semaphores are found to be in an invalid state.

--Woof!


git-svn-id: https://sipxecs.sipfoundry.org/rep/sipXecs/main@12049 ab1d8caa-1f67-47f1-9e81-24633a41865c