XCF-2347 fix switching location on Services page in HA mode
Add listener for Services form: it cleans the current services status which forces the page to reload the new status.
It fixes the problem of displaying the old status after location has been changed.
Unfortunately in some cases (refresh) we retrieve the services status twice (in rewind and in render phases), but it does not seem to affect user experience.


git-svn-id: https://sipxecs.sipfoundry.org/rep/sipXecs/main@12052 ab1d8caa-1f67-47f1-9e81-24633a41865c