Changed method to detect hung IMDB for read/write/update semaphores
As the state of the database lock is held in shared memory,
and not by the semaphores itself, checking the semaphore value
for "0" doesn't work, as it will be 0 even in the "locked" state,
as this semaphore is actually used as a trigger to unblock
waiters who only wait if the shared memory tells them a lock is
held.

The best way to detect a hung IMDB on these locks is to check
that the semncnt (number of threads waiting for the value
to increase) is not 0 for a long time.  If the semncnt is 0,
no one is waiting on the lock.  If it is greater than 0,
that many threads are waiting.  The number of waiters should
be 0 after a short amount of time.

--Woof!


git-svn-id: https://sipxecs.sipfoundry.org/rep/sipXecs/main@12109 ab1d8caa-1f67-47f1-9e81-24633a41865c