XECS-265 - Support RFC 3327 (Path header)
Patch from Robert Joly.  Add code to support RFC 3327 Path Header 
Changes allow the Registry code to recognize path headers in a Register request,
store them and return them in the 200 response. Nothing is added to use them
in the proxy or other services.

IMPORTANT NOTE (especially for developers):
This code adds a column to the imdb.  Unfortunately, the SipRouterTest.cpp unit 
test was using the "real" database file  (...sipx/var/sipxdata/tmp/imdb.odb).
Consequently, if you update and run code with this fix, an attempt to
build any previous versions of the code base will fail this unit test.  This
includes 3.8 and 3.10 versions.  The error message says it can't delete
because it isn't empty.
The simplest work around is to delete the ...sipx/var/sipxdata/tmp/imdb.odb
file whenever you are going to build an earlier version.  
The SipRouterTest.cpp has been corrected to use a test db file in this version.
That correction could be back-merged into 3.10 and 3.8 to eliminate this
annoyance.
<ke>


git-svn-id: https://sipxecs.sipfoundry.org/rep/sipXecs/main@12116 ab1d8caa-1f67-47f1-9e81-24633a41865c