XCF-1781 Tapestry upgrade - do not render empty If/Else statements
Tapestry 4.1 will by default render tags for If, Any, Else, For elements.  In many cases it's what
we want, but sometimes - especially when the component is question is inside of <span /> or <div />
we do not want to render it.  Rendering can be switched off by setting renderTag parameter to
'false'.


git-svn-id: https://sipxecs.sipfoundry.org/rep/sipXecs/main@12300 ab1d8caa-1f67-47f1-9e81-24633a41865c