XCF-2370 move linksys phone and ata settigns file to a single linksys directory
Each plugin only support generating a single .jar file and a single settings directory.
Linksys plugin used to kept phone settings in linksysPhone, and ATA settings in linksysATA
Now all settings files for both ATAs and Phones are in linksys directory.

Keeping all the files in a single directory required a new capability of the PhoneModel: it now has
two new properties one can specify the base name of the settings file for a device and for a line
(defaults are 'phone.xml' and 'line.xml' which used to be hardcoded in a Phone.java.


git-svn-id: https://sipxecs.sipfoundry.org/rep/sipXecs/main@12380 ab1d8caa-1f67-47f1-9e81-24633a41865c