XCF-1772 linksys: copy default model-specific configuration file during profile generation
Linksys phones by default look for a minimum initial configuration file, which in turns provide the
name of the device specific configuration file.
The name of the file seems to depend on the device model. Most documents on the web suggest it's
'spa$PSN.cfg' where $PSN refers to numeric portion of the model name.
LinksysModel builds the name (from 'psn' property set in model definition) and Linksys.copyFiles is
overwritten to copy default.cfg template into the file with appropriate name.


git-svn-id: https://sipxecs.sipfoundry.org/rep/sipXecs/main@12381 ab1d8caa-1f67-47f1-9e81-24633a41865c