XCF-2602 Implement XML/RPC timeout.
This is probably not the best solution: just a workaround for the XML/RPC library limitation.
As far as I can see the default transport provider (based on standard JVM HTTP client) is configured
with infinite timeout and there is no way to change it. Until we figure that out I added simple
timeout implementation in XmlRpcClientInterceptor (one more reason to use it).  We call XML/RPC
library in background thread and timeout if do not get response. The default timeout is 5 seconds,
configurable in milliseconds through standard Spring properties.


git-svn-id: https://sipxecs.sipfoundry.org/rep/sipXecs/main@12890 ab1d8caa-1f67-47f1-9e81-24633a41865c